include/ta_fiovb.h: add hsm_so_pin variable
This patch is part of a set of patches across multiple repositories in
order to add support from cleanup of the RPMB in case of a full factory
reset.

When preforming a full factory reset, all objects with regards to aktualizr-lite
should be properly cleared from the RPMB storage in order to ensure that
no data is left. This can be achieved by re-running the PKCS11
`InitToken()` function.

In order to perform the re-initialisation, we need the SO pin which was
used to initialize the token in the first place. Therefor, we create a
`hsm_so_pin` variable in fiovb. This variable will be set by the
`lmp-device-register` script via the `fiovb_setenv` and read during the
factory reset using `fiovb_printenv`.

Signed-off-by: Bram Vlerick <bram.vlerick@openpixelsystems.org>